# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Traffic Peak typing platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /certificates:
    get:
      operationId: listCertificates
      tags: [certificates]
      summary: List certificates for current user
      responses:
        "200":
          description: Certificates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Certificate"
    post:
      operationId: generateCertificate
      tags: [certificates]
      summary: Generate a certificate for a test result
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CertificateInput"
      responses:
        "201":
          description: Certificate generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Certificate"
  /certificates/check-eligibility:
    post:
      operationId: checkCertificateEligibility
      tags: [certificates]
      summary: Check if a test result is eligible for a certificate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                testResultId:
                  type: integer
              required: [testResultId]
      responses:
        "200":
          description: Eligibility check result
          content:
            application/json:
              schema:
                type: object
                properties:
                  eligible:
                    type: boolean
                  errors:
                    type: array
                    items:
                      type: string
                  reasons:
                    type: object
                    additionalProperties:
                      type: string
                  stats:
                    type: object
                    properties:
                      wpm:
                        type: number
                      accuracy:
                        type: number
                      duration:
                        type: integer
                      minWpm:
                        type: integer
                      minDuration:
                        type: integer
  /certificates/{certificateId}:
    get:
      operationId: verifyCertificate
      tags: [certificates]
      summary: Verify a certificate by ID (public)
      parameters:
        - name: certificateId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Certificate is valid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Certificate"
        "404":
          description: Certificate not found

  # ── Blog ──────────────────────────────────────────────────────────────
...
  /assessment/{token}:
    get:
      operationId: getAssessmentByToken
      tags: [assessment]
      summary: Get public assessment info (no auth)
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Assessment info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicAssessment"
        "404":
          description: Not found or expired
  /assessment/{token}/start:
    post:
      operationId: startAssessment
      tags: [assessment]
      summary: Register candidate and start session
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CandidateIntakeInput"
      responses:
        "201":
          description: Candidate session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CandidateSession"
        "400":
          description: Already attempted or validation error
        "404":
          description: Assessment not found
  /assessment/{token}/submit:
    post:
      operationId: submitAssessment
      tags: [assessment]
      summary: Submit assessment result
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssessmentSubmitInput"
      responses:
        "200":
          description: Result saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssessmentSubmitResult"
        "400":
          description: Validation error
        "404":
          description: Not found
components:
  schemas:
...
    CertificateInput:
      type: object
      required: [testResultId, recipientName]
      properties:
        testResultId:
          type: integer
        recipientName:
          type: string
    Certificate:
      type: object
      required: [id, certificateId, recipientName, wpm, accuracy, duration, mode, level, issuedAt]
      properties:
        id:
          type: integer
        certificateId:
          type: string
        userId:
          type: ["integer", "null"]
        testResultId:
          type: ["integer", "null"]
        recipientName:
          type: string
        wpm:
          type: number
        accuracy:
          type: number
        duration:
          type: integer
        mode:
          type: string
        level:
          type: string
          enum: [Novice, Beginner, Intermediate, Advanced]
        issuedAt:
          type: string

    # ── Blog ──────────────────────────────────────────────────────────
...
    PublicAssessment:
      type: object
      required: [id, token, name, companyName, jobPosition, durationSeconds, passingWpm, minAccuracy, maxAttempts]
      properties:
        id:
          type: integer
        token:
          type: string
        name:
          type: string
        companyName:
          type: string
        jobPosition:
          type: string
        description:
          type: ["string", "null"]
        durationSeconds:
          type: integer
        difficulty:
          type: string
        language:
          type: string
        contentType:
          type: string
        customText:
          type: ["string", "null"]
        passingWpm:
          type: integer
        minAccuracy:
          type: number
        maxAttempts:
          type: integer
        expiresAt:
          type: ["string", "null"]
    CandidateIntakeInput:
      type: object
      required: [fullName, email]
      properties:
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
        fingerprint:
          type: string
    CandidateSession:
      type: object
      required: [candidateId, assessmentId, durationSeconds, passingWpm, minAccuracy]
      properties:
        candidateId:
          type: integer
        assessmentId:
          type: integer
        durationSeconds:
          type: integer
        contentType:
          type: string
        customText:
          type: ["string", "null"]
        passingWpm:
          type: integer
        minAccuracy:
          type: number
    AssessmentSubmitInput:
      type: object
      required: [candidateId, wpm, cpm, accuracy, errorCount, charCount, durationSeconds]
      properties:
        candidateId:
          type: integer
        wpm:
          type: number
        cpm:
          type: number
        accuracy:
          type: number
        errorCount:
          type: integer
        charCount:
          type: integer
        durationSeconds:
          type: integer
        tabSwitches:
          type: integer
    AssessmentSubmitResult:
      type: object
      required: [id, wpm, cpm, accuracy, errorCount, passed, passingWpm, minAccuracy]
      properties:
        id:
          type: integer
        wpm:
          type: number
        cpm:
          type: number
        accuracy:
          type: number
        errorCount:
          type: integer
        passed:
          type: boolean
        passingWpm:
          type: integer
        minAccuracy:
          type: number
        companyName:
          type: string
        jobPosition:
          type: string